move EnvironmentInfo and GitEnvironmentInfo classes to model folder
