+++ src/wrappers/foresight/model/GitEnvironmentInfo.ts
